+++ complex/library/includes/classes/class-updater.php
@@ -7,9 +7,6 @@
  *
  * @package     WebMan WordPress Theme Framework
  * @subpackage  Updater
- *
- * @since    1.0
- * @version  1.3
  */